+++ admin/views/media/index.php
@@ -78,7 +78,7 @@
                                                 <td><?php echo $media['createDate']; ?></td>
                                                 <td><?php echo pathinfo('public/upload/media/' . $media['slug'], PATHINFO_EXTENSION); ?></td>
                                                 <td><a href="admin.php?controller=media&amp;action=edit&amp;media_id=<?php echo $media['id']; ?>" class="btn btn-default waves-effect waves-float btn-sm waves-green"><i class="zmdi zmdi-edit"></i></a>
-                                                    <a href="admin.php?controller=media&amp;action=delete&amp;media_id=<?php echo $media['id']; ?>" class="btn btn-default waves-effect waves-float btn-sm waves-red"><i class="zmdi zmdi-delete"></i></a>
+                                                    <a onclick="return confirm('Are you sure to delete?')" href="admin.php?controller=media&amp;action=delete&amp;media_id=<?php echo $media['id']; ?>" class="btn btn-default waves-effect waves-float btn-sm waves-red"><i class="zmdi zmdi-delete"></i></a>
                                                     <a href="public/upload/media/<?php echo $media['slug']; ?>" target="_blank" class="btn btn-default waves-effect waves-float btn-sm waves-red"><i class="zmdi zmdi-square-right"></i></a></td>
                                             </tr>
                                         <?php endforeach; ?>
