+++ admin/views/shop/index.php
@@ -72,7 +72,7 @@
                                                 <td><?php echo $category['slug'] ?></td>
                                                 <td><?php echo $category['category_position'] ?></td>
                                                 <td><a href="admin.php?controller=shop&amp;action=edit&amp;cate_id=<?php echo $category['id']; ?>" class="btn btn-default waves-effect waves-float btn-sm waves-green"><i class="zmdi zmdi-edit"></i></a>
-                                                    <a href="admin.php?controller=shop&amp;action=delete&amp;cate_id=<?= $category['id'] ?>" class="btn btn-default waves-effect waves-float btn-sm waves-red"><i class="zmdi zmdi-delete"></i></a></td>
+                                                    <a onclick="return confirm('Are you sure to delete?')" href="admin.php?controller=shop&amp;action=delete&amp;cate_id=<?= $category['id'] ?>" class="btn btn-default waves-effect waves-float btn-sm waves-red"><i class="zmdi zmdi-delete"></i></a></td>
                                             </tr>
                                         <?php endforeach; ?>
                                     </tbody>
